+++ examples/resources/azexecute_application_request/resource.tf
@@ -4,7 +4,7 @@ terraform {
   required_providers {
     azexecute = {
       source  = "dyntora/azexecute"
-      version = "~> 0.5"
+      version = "~> 0.6"
     }
   }
 }